Install UltiSnips, Supertab, compatible with YCM
Displays suggestions from US and YCM in the same list.

Now uses <C-n> and <C-p> for YCM and <Tab> and <S-Tab> for US,
unless there's no snip in which case tab for YCM too.

A little bit messy - if you type something and both YCM and US
something, but YCM's is at the top of the list, <Tab> will insert the
snippet immediate. Not perfect, but it's pretty damn good anyway.