Snippet 'asa' defines contained text better
Using ${0} instead of ${1} means that UltiSnips will stop being
interested once ${0} is reached.

The effect I've noticed so far is that I can use <Tab> inside the block
(for indenting etc) and it will affect the entered text. Prior to this
commit, using <Tab> would jump to the end of the snippet, i.e. after
\end{lstlisting}.