Switch from Vundle to Vim-Plug
Concurrency! And experiencing new things!

Vundle in this repo was a Git submodule living in .vim. Vim-Plug has
been stored as a static file, as suggested by its owner.
As a result, it will occasionally need updating.
(it can update itself with PlugUpgrade but that still needs committing)

I'm willing to live with that for now.