Use bat -p instead of vp for default PAGER
`vp` is an alias, which means eg `git log` doesn't know what it is.
You can probably fix this by making it a shell script, but bat is meant
for this kind of thing anyway.

I'll run them in parallel now and see which one I prefer. 'bat' will
trigger on default stuff using PAGER and 'vp' will be manually invoked,
likely in longer command pipelines.